Switch to using pip-compile to manage requirements
pip-compile is from the future branch of pip-tools, and allows us to
specify in *requirements.in files the top-level dependencies of our
project, which are then compiled to *requirements.txt files containing
all dependencies. This makes it easier to keep track of:

- which libraries we're using directly
- which are pulled in as subdependencies
- which ones we actually care about the version of

To add a new dependency, add a line to the relevant *requirements.in
file and then run `pip-compile`. Then `pip install -r <relevant requirements.txt>`
to install it.