Parse bearer token out of Authorization header
This sets the WWW-Authenticate header on the response following
RFC7235 [1], but doesn't actually make a request to Signon yet,
instead assuming that a provided token is always valid.

Since we now have a test dependency (mock) which isn't implicitly
in the development requirements already (as nose is, via sniffer),
start including test-requirements.in in development-requirements.in.

[1] https://tools.ietf.org/html/rfc7235#section-3.1