Move request logging into middleware
This now logs something for every request and response, rather than
only successful PUT requests.

There's a big problem with this, though: error responses generated
by raising HTTPError subclasses are logged as 200 instead of the
actual response status code. This affects the following cases:

- missing/invalid bearer token
- wrong Content-Type header
- wrong Accept header
- PUT data fails validation against the JSON schema

Falcon calls middleware on the response before handling HTTPError
subclasses and modifying the response based on them. That means that
at the point that the response is logged, the status is still the
default 200.

Error responses due to no route being matched or method not allowed
are correctly logged, because Falcon modifies the response without
raising an HTTPError in these cases.