Send a valid court to the Publishing API
If the status code from the publishing API isn't 200/201, or if
requests.ConnectionError is raised, this returns a standard 503 response.
We should probably return a more informative message in some of those
cases.

The tests so far only assert that a PUT request is made to the publishing
API, and not that the correct data is sent. We should validate the
data against a schema in govuk_content_schemas when we have added one
for the court format.