Convert logging into WSGI middleware
Logging of all requests was previously implemented using Falcon's
middleware interface, but that was unable to log the response status
accurately when Falcon's error classes were raised. Moving this logging
outside the application and into WSGI middleware solves this problem.

We have to do a couple of potentially odd-looking things here in order
to gain access to fields we want to log:

* The request body: the body isn't normally read outside the application,
  and it's a stream (file-like object), so in order to have it available
  to log we need to read it in middleware, but then replace it in the
  environ dict with a new file-like object so that it can be read again
  in the application. With Gunicorn, the wsgi.input stream doesn't have
  a seek() method, so we can't rewind the same object after reading it.

* The response status code: the status is normally passed directly to the
  start_response callable from the application and so no reference to it
  exists at the middleware level. By wrapping the start_response callable,
  we can keep a reference to the status in order to log it.

The entire request body is read before authentication takes place, so it
may be possible for a malicious user to make very large requests which
allocate all available memory, but hpoefully nginx will prevent that.